fix: mvp용 기능 수정
