+++ .gitignore
@@ -1,0 +1,8 @@
+# .env 파일 무시
+v1_fix/.env
+
+# __pycache__ 디렉토리 무시
+v1_fix/__pycache__/
+
+# 가상환경 디렉토리 무시
+v1_fix/marong
