--- requirements.txt
@@ -1,155 +1,0 @@
-accelerate==1.6.0
-annotated-types==0.7.0
-anyio==4.9.0
-appnope==0.1.4
-asgiref==3.8.1
-asttokens==3.0.0
-attrs==25.3.0
-backoff==2.2.1
-bcrypt==4.3.0
-build==1.2.2.post1
-cachetools==5.5.2
-certifi==2025.4.26
-charset-normalizer==3.4.1
-chroma-hnswlib==0.7.6
-chromadb==1.0.7
-click==8.1.8
-coloredlogs==15.0.1
-comm==0.2.2
-debugpy==1.8.14
-decorator==5.2.1
-Deprecated==1.2.18
-distro==1.9.0
-dnspython==2.7.0
-durationpy==0.9
-email_validator==2.2.0
-emoji==2.14.1
-executing==2.2.0
-fastapi==0.115.9
-fastapi-cli==0.0.7
-filelock==3.18.0
-flatbuffers==25.2.10
-fsspec==2025.3.2
-google-auth==2.39.0
-googleapis-common-protos==1.70.0
-grpcio==1.71.0
-h11==0.16.0
-httpcore==1.0.9
-httptools==0.6.4
-httpx==0.28.1
-huggingface-hub==0.30.2
-humanfriendly==10.0
-idna==3.10
-importlib_metadata==8.6.1
-importlib_resources==6.5.2
-ipykernel==6.29.5
-ipython==8.22.2
-ipython_pygments_lexers==1.1.1
-jedi==0.19.2
-Jinja2==3.1.6
-joblib==1.4.2
-jsonpatch==1.33
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2025.4.1
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-kubernetes==32.0.1
-langchain==0.3.24
-langchain-core==0.3.56
-langchain-text-splitters==0.3.8
-langsmith==0.3.38
-markdown-it-py==3.0.0
-MarkupSafe==3.0.2
-matplotlib-inline==0.1.7
-mdurl==0.1.2
-mmh3==5.1.0
-monotonic==1.6
-mpmath==1.3.0
-nest-asyncio==1.6.0
-networkx==3.4.2
-numpy==2.2.5
-oauthlib==3.2.2
-onnxruntime==1.21.1
-opentelemetry-api==1.32.1
-opentelemetry-exporter-otlp-proto-common==1.32.1
-opentelemetry-exporter-otlp-proto-grpc==1.32.1
-opentelemetry-instrumentation==0.53b1
-opentelemetry-instrumentation-asgi==0.53b1
-opentelemetry-instrumentation-fastapi==0.53b1
-opentelemetry-proto==1.32.1
-opentelemetry-sdk==1.32.1
-opentelemetry-semantic-conventions==0.53b1
-opentelemetry-util-http==0.53b1
-orjson==3.10.16
-overrides==7.7.0
-packaging==24.2
-pandas==2.2.3
-parso==0.8.4
-pexpect==4.9.0
-pillow==11.2.1
-platformdirs==4.3.7
-posthog==4.0.0
-prompt_toolkit==3.0.51
-protobuf==5.29.4
-psutil==7.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pyasn1==0.6.1
-pyasn1_modules==0.4.2
-pydantic==2.11.3
-pydantic_core==2.33.1
-Pygments==2.19.1
-PyMySQL==1.1.1
-PyPika==0.48.9
-pyproject_hooks==1.2.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
-python-multipart==0.0.20
-pytz==2025.2
-PyYAML==6.0.2
-pyzmq==26.4.0
-referencing==0.36.2
-regex==2024.11.6
-requests==2.32.3
-requests-oauthlib==2.0.0
-requests-toolbelt==1.0.0
-rich==14.0.0
-rich-toolkit==0.14.4
-rpds-py==0.24.0
-rsa==4.9.1
-safetensors==0.5.3
-scikit-learn==1.6.1
-scipy==1.15.2
-sentence-transformers==4.1.0
-setuptools==75.8.0
-shellingham==1.5.4
-six==1.17.0
-sniffio==1.3.1
-SQLAlchemy==2.0.40
-stack-data==0.6.3
-starlette==0.45.3
-sympy==1.14.0
-tenacity==9.1.2
-threadpoolctl==3.6.0
-tokenizers==0.21.1
-torch==2.7.0
-tornado==6.4.2
-tqdm==4.67.1
-traitlets==5.14.3
-transformers==4.51.3
-typer==0.15.3
-typing-inspection==0.4.0
-typing_extensions==4.13.2
-tzdata==2025.2
-urllib3==2.4.0
-uvicorn==0.34.2
-uvloop==0.21.0
-watchfiles==1.0.5
-wcwidth==0.2.13
-websocket-client==1.8.0
-websockets==15.0.1
-wheel==0.45.1
-wrapt==1.17.2
-zipp==3.21.0
-zstandard==0.23.0
