--- run_chroma.sh
@@ -1,7 +1,0 @@
-# run_chroma.sh (예: 프로젝트 루트에 저장)
-#!/bin/bash
-
-chroma run \
-  --host 0.0.0.0 \
-  --port 8001 \
-  --path /Users/kimss/Documents/marong/chroma_db
