--- template.md
@@ -1,43 +1,0 @@
-## Commit Message Template
-
-################
-
-# <타입> : <제목> 의 형식으로 제목을 아래 공백줄에 작성
-
-# 제목은 50자 이내 / 변경사항이 "무엇"인지 명확히 작성 / 끝에 마침표 금지 / 명령어로 작성
-
-# 예) feat : 로그인 기능 추가
-
-# 바로 아래 공백은 지우지 마세요 (제목과 본문의 분리를 위함)
-
-################
-
-# 본문(구체적인 내용)을 아랫줄에 작성
-
-# 여러 줄의 메시지를 작성할 땐 "-"로 구분 (한 줄은 72자 이내)
-
-################
-
-# 꼬릿말(footer)을 아랫줄에 작성 (현재 커밋과 관련된 이슈 번호 추가 등)
-
-# 예) #7
-
-################
-
-# feat : 새로운 기능 추가
-
-# fix : 버그 수정
-
-# docs : 문서 수정
-
-# test : 테스트 코드 추가
-
-# refactor : 코드 리팩토링
-
-# style : 코드 포맷팅, 세미콜론 누락 등 코드 변경이 없는 경우
-
-# chore : 빌드, 패키지 매니저 수정
-
-# deploy: 배포 CI/CD
-
-################
