+++ v1_fix/sbert_down.py
@@ -1,0 +1,3 @@
+from sentence_transformers import SentenceTransformer
+model = SentenceTransformer("snunlp/KR-SBERT-V40K-klueNLI-augSTS")
+model.save("./kr-sbert")
